things are moving.
the webui is working, backups are working, disk detection is next

TODO: an interface where a user can specify that a disk should be a source /
a target

Also, maybe an interface for unmounting those disks.

I'm feeling a `/disks` view coming along.

Then, all we need is a nice GUI button for 'shutdown'